Changed highest frequency from 1.624 to 1.7GHz
Tweaked DVFS table for the GPU. It should now scale a bit better and
still bring the same performance and the top end.

Increased core voltage for the highest frequency to 1250mV.

Lowest brightness setting increased from 13 to 18 (thanks to
clemsyn). Lets give this a try and we can increase it further if need
be.